Refactors for the feedback feature. Moved under user_feedback folder. Form Display disabled till DB gets in place
