+++ manifest.json
@@ -1,10 +1,10 @@
 {
     "manifest_version": 3,
     "name": "Better My Email",
-    "version": "2.1",
+    "version": "2.2",
     "web_accessible_resources": [
         {
-            "resources": ["dist/client/feedbackForm.html"],
+            "resources": ["dist/client/user_feedback/feedbackForm.html"],
             "matches": ["*://mail.google.com/*"]
         }
     ],
