+++ package.json
@@ -6,7 +6,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "build:client": "webpack --config webpack.config.cjs",
     "build:server": "tsc -p tsconfig.server.json",
-    "copy:feedbackForm": "cp src/client/feedbackForm.html dist/client/feedbackForm.html",
+    "copy:feedbackForm": "cp src/client/user_feedback/feedbackForm.html dist/client/user_feedback/feedbackForm.html",
     "build": "npm run build:client && npm run build:server && npm run copy:feedbackForm",
     "start": "node dist/server/apiGatewayServer.js"
   },
